requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pillow-avif>=1.3.0
pillow-heif>=0.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
duckduckgo-search>=3.9.0
tqdm>=4.65.0 